Add support for data attribute instead of file
