+++ public/tracks.js
@@ -1,0 +1,100 @@
+export default [
+  {
+    file: "https://res.cloudinary.com/dpmchqezv/video/upload/v1631305734/troy/alone-in-a-crowded-room/01_I_Know_I_m_Not_jvp3af.mp3",
+    title: "I Know I'm Not",
+    artist: "troy.",
+    artistUrl: "https://troyalllowercase.bandcamp.com/",
+    album: "Alone In A Crowded Room",
+    albumUrl: "https://troyalllowercase.bandcamp.com/album/alone-in-a-crowded-room",
+    artwork:
+      "https://res.cloudinary.com/dpmchqezv/image/upload/c_scale,f_auto,q_auto:eco,w_300/v1631306010/troy/alone-in-a-crowded-room/AIACR-troy_cwoizg.png",
+  },
+  {
+    file: "https://res.cloudinary.com/dpmchqezv/video/upload/v1631305724/troy/alone-in-a-crowded-room/02_Baby_Blue_lwf7zf.mp3",
+    title: "Baby Blue",
+    artist: "troy.",
+    album: "Alone In A Crowded Room",
+    artwork:
+      "https://res.cloudinary.com/dpmchqezv/image/upload/c_scale,f_auto,q_auto:eco,w_300/v1631306010/troy/alone-in-a-crowded-room/AIACR-troy_cwoizg.png",
+  },
+  {
+    file: "https://res.cloudinary.com/dpmchqezv/video/upload/v1631305729/troy/alone-in-a-crowded-room/03_Insincerity_x29pxf.mp3",
+    title: "Insincerity",
+    artist: "troy.",
+    album: "Alone In A Crowded Room",
+    artwork:
+      "https://res.cloudinary.com/dpmchqezv/image/upload/c_scale,f_auto,q_auto:eco,w_300/v1631306010/troy/alone-in-a-crowded-room/AIACR-troy_cwoizg.png",
+  },
+  {
+    file: "https://res.cloudinary.com/dpmchqezv/video/upload/v1631305723/troy/alone-in-a-crowded-room/04_Like_Sinking_ginqay.mp3",
+    title: "Like Sinking",
+    artist: "troy.",
+    album: "Alone In A Crowded Room",
+    artwork:
+      "https://res.cloudinary.com/dpmchqezv/image/upload/c_scale,f_auto,q_auto:eco,w_300/v1631306010/troy/alone-in-a-crowded-room/AIACR-troy_cwoizg.png",
+  },
+  {
+    file: "https://res.cloudinary.com/dpmchqezv/video/upload/v1631305726/troy/alone-in-a-crowded-room/05_Do_Better_fdvpxz.mp3",
+    title: "Do Better",
+    artist: "troy.",
+    album: "Alone In A Crowded Room",
+    artwork:
+      "https://res.cloudinary.com/dpmchqezv/image/upload/c_scale,f_auto,q_auto:eco,w_300/v1631306010/troy/alone-in-a-crowded-room/AIACR-troy_cwoizg.png",
+  },
+  {
+    file: "https://res.cloudinary.com/dpmchqezv/video/upload/v1631305732/troy/alone-in-a-crowded-room/06_Market_Street_nrdgi4.mp3",
+    title: "Market Street",
+    artist: "troy.",
+    album: "Alone In A Crowded Room",
+    artwork:
+      "https://res.cloudinary.com/dpmchqezv/image/upload/c_scale,f_auto,q_auto:eco,w_300/v1631306010/troy/alone-in-a-crowded-room/AIACR-troy_cwoizg.png",
+  },
+  {
+    file: "https://res.cloudinary.com/dpmchqezv/video/upload/v1631305735/troy/alone-in-a-crowded-room/07_Alone_In_a_Crowded_Room_zyrerb.mp3",
+    title: "Alone In A Crowded Room",
+    artist: "troy.",
+    album: "Alone In A Crowded Room",
+    artwork:
+      "https://res.cloudinary.com/dpmchqezv/image/upload/c_scale,f_auto,q_auto:eco,w_300/v1631306010/troy/alone-in-a-crowded-room/AIACR-troy_cwoizg.png",
+  },
+  {
+    file: "https://res.cloudinary.com/dpmchqezv/video/upload/v1631305740/troy/alone-in-a-crowded-room/08_Long_Time_Caller_First_Time_List_plvy7c.mp3",
+    title: "Long Time Caller, First Time Listener",
+    artist: "troy.",
+    album: "Alone In A Crowded Room",
+    artwork:
+      "https://res.cloudinary.com/dpmchqezv/image/upload/c_scale,f_auto,q_auto:eco,w_300/v1631306010/troy/alone-in-a-crowded-room/AIACR-troy_cwoizg.png",
+  },
+  {
+    file: "https://res.cloudinary.com/dpmchqezv/video/upload/v1631305733/troy/alone-in-a-crowded-room/09_Detroit_uvrmff.mp3",
+    title: "Detroit",
+    artist: "troy.",
+    album: "Alone In A Crowded Room",
+    artwork:
+      "https://res.cloudinary.com/dpmchqezv/image/upload/c_scale,f_auto,q_auto:eco,w_300/v1631306010/troy/alone-in-a-crowded-room/AIACR-troy_cwoizg.png",
+  },
+  {
+    file: "https://res.cloudinary.com/dpmchqezv/video/upload/v1631305740/troy/alone-in-a-crowded-room/10_We_Can_t_Rush_These_Things_mhhr0b.mp3",
+    title: "We Can't Rush These Things",
+    artist: "troy.",
+    album: "Alone In A Crowded Room",
+    artwork:
+      "https://res.cloudinary.com/dpmchqezv/image/upload/c_scale,f_auto,q_auto:eco,w_300/v1631306010/troy/alone-in-a-crowded-room/AIACR-troy_cwoizg.png",
+  },
+  {
+    file: "https://res.cloudinary.com/dpmchqezv/video/upload/v1631305741/troy/alone-in-a-crowded-room/11_Closure_ttfryn.mp3",
+    title: "Closure",
+    artist: "troy.",
+    album: "Alone In A Crowded Room",
+    artwork:
+      "https://res.cloudinary.com/dpmchqezv/image/upload/c_scale,f_auto,q_auto:eco,w_300/v1631306010/troy/alone-in-a-crowded-room/AIACR-troy_cwoizg.png",
+  },
+  {
+    file: "https://res.cloudinary.com/dpmchqezv/video/upload/v1631305740/troy/alone-in-a-crowded-room/12_Like_Drowning_Bonus_Track_ta2ro5.mp3",
+    title: "Like Drowning (Bonus Track)",
+    artist: "troy.",
+    album: "Alone In A Crowded Room",
+    artwork:
+      "https://res.cloudinary.com/dpmchqezv/image/upload/c_scale,f_auto,q_auto:eco,w_300/v1631306010/troy/alone-in-a-crowded-room/AIACR-troy_cwoizg.png",
+  },
+];
